parametros
