+++ src/parametros/crud-ceco/ceco.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from '@nestjs/common';
+
+@Controller('ceco')
+export class CecoController { }
