+++ src/parametros/crud-cliente/cliente.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from '@nestjs/common';
+
+@Controller('cliente')
+export class ClienteController {}
