+++ src/parametros/crud-estados/estados.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from "@nestjs/common";
+
+@Controller('estados')
+export class EstadosController {}
